requirements.txt
openai>=1.12.0
supabase>=2.0.0
requests>=2.31.0
selenium>=4.18.0
python-dotenv>=1.0.1
